requirements.txt
python-telegram-bot==21.5
python-dotenv==1.0.1
certifi==2024.8.30
